pyproject.toml
[project]
name = "sregym"
requires-python = ">= 3.12"
version = "0.0.1"

dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.1",
    "aiosignal==1.3.2",
    "airportsdata==20250523",
    "annotated-types==0.7.0",
    "ansible>=11.8.0",
    "anthropic>=0.52.0",
    "anyio==4.9.0",
    "apscheduler>=3.11.0",
    "asgiref==3.8.1",
    "astor==0.8.1",
    "attrs==25.3.0",
    "autogen-agentchat==0.2.40",
    "azure-ai-ml==1.27.1",
    "azure-common==1.1.28",
    "azure-core==1.34.0",
    "azure-core-tracing-opentelemetry==1.0.0b12",
    "azure-identity==1.23.0",
    "azure-mgmt-core==1.5.0",
    "azure-monitor-opentelemetry==1.6.9",
    "azure-monitor-opentelemetry-exporter==1.0.0b36",
    "azure-storage-blob==12.25.1",
    "azure-storage-file-datalake==12.20.0",
    "azure-storage-file-share==12.21.0",
    "bashlex>=0.18",
    "bcrypt==4.3.0",
    "black==24.10.0",
    "blake3==1.0.5",
    "boto3>=1.28.57",
    "bs4>=0.0.2",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "cloudpickle==3.1.1",
    "colorama==0.4.6",
    "compressed-tensors==0.9.1",
    "contourpy==1.3.2",
    "cryptography==45.0.3",
    "cycler==0.12.1",
    "dateparser==1.2.1",
    "deprecated==1.2.18",
    "depyf==0.18.0",
    "dill==0.4.0",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docker==7.1.0",
    "docker-pycreds==0.4.0",
    "einops==0.8.1",
    "elastic-transport==8.17.1",
    "elasticsearch==8.18.1",
    "email-validator==2.2.0",
    "fastapi[standard]==0.115.12",
    "fastapi-cli==0.0.7",
    "fastmcp>=2.9.2",
    "filelock==3.18.0",
    "fixedint==0.1.6",
    "flaml==2.3.4",
    "fonttools==4.58.0",
    "frozenlist==1.6.0",
    "fsspec==2025.5.1",
    "geni-lib-xlab",
    "gguf==0.10.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.40.2",
    "h11==0.16.0",
    "hf-xet==1.1.2",
    "httmock==1.4.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "huggingface-hub==0.32.1",
    "idna==3.10",
    "importlib==1.0.4",
    "importlib-metadata==8.6.1",
    "interegular==0.3.3",
    "isodate==0.7.2",
    "itsdangerous>=2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "kiwisolver==1.4.8",
    "kubernetes==30.1.0",
    "langchain>=0.3.25",
    "langchain-litellm>=0.2.2",
    "langchain-openai>=0.3.18",
    "langgraph>=0.4.7",
    "langsmith>=0.3.43",
    "lark==1.2.2",
    "litellm>=1.71.1",
    "llvmlite==0.43.0",
    "lm-format-enforcer==0.10.11",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib==3.10.3",
    "mcp[cli]>=1.9.1",
    "mdurl==0.1.2",
    "mistral-common==1.5.6",
    "mpmath==1.3.0",
    "msal==1.32.3",
    "msal-extensions==1.3.1",
    "msgpack==1.1.0",
    "msgspec==0.19.0",
    "msrest==0.7.1",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nodeenv==1.9.1",
    "numba==0.60.0",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "opentelemetry-api==1.31.1",
    "opentelemetry-instrumentation==0.52b1",
    "opentelemetry-instrumentation-asgi==0.52b1",
    "opentelemetry-instrumentation-dbapi==0.52b1",
    "opentelemetry-instrumentation-django==0.52b1",
    "opentelemetry-instrumentation-fastapi==0.52b1",
    "opentelemetry-instrumentation-flask==0.52b1",
    "opentelemetry-instrumentation-psycopg2==0.52b1",
    "opentelemetry-instrumentation-requests==0.52b1",
    "opentelemetry-instrumentation-urllib==0.52b1",
    "opentelemetry-instrumentation-urllib3==0.52b1",
    "opentelemetry-instrumentation-wsgi==0.52b1",
    "opentelemetry-resource-detector-azure==0.1.5",
    "opentelemetry-sdk==1.31.1",
    "opentelemetry-semantic-conventions==0.52b1",
    "opentelemetry-util-http==0.52b1",
    "outlines==0.1.11",
    "outlines-core==0.1.26",
    "packaging>=24.0",
    "pandas==2.2.3",
    "paramiko==3.5.1",
    "partial-json-parser==0.2.1.1.post5",
    "pathspec==0.12.1",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "pre-commit>=4.2.0",
    "prometheus-api-client==0.5.7",
    "prometheus-client==0.22.0",
    "prometheus-fastapi-instrumentator==7.1.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "protobuf==6.31.0",
    "psutil==6.1.1",
    "py-cpuinfo==9.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycountry==24.6.1",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydash==8.0.5",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pynacl==1.5.0",
    "pyparsing==3.2.3",
    "pyright==1.1.401",
    "pytest>=8.3.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "ray==2.40.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests>=2.32.3",
    "requests-oauthlib==2.0.0",
    "rich==13.9.4",
    "rich-toolkit==0.14.6",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "sentencepiece==0.2.0",
    "sentry-sdk==2.29.1",
    "setproctitle==1.3.6",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "strictyaml==1.7.3",
    "sympy==1.13.1",
    "termcolor==3.1.0",
    "tiktoken==0.7.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "transformers==4.52.3",
    "typer==0.16.0",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "urllib3>=2.4.0",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    "watchfiles==1.0.5",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.20.0",
    "zipp==3.22.0",
    "pyfiglet>=1.0.3",
    "langchain-ibm>=0.3.15",
    "langchain-google-genai>=2.1.12",
    "dash>=3.2.0",
]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["setuptools>=80.8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["sregym", "clients", "provisioner", "scripts"]

[tool.uv.sources]
geni-lib-xlab = { path = "scripts/geni_lib/mod/geni_lib_xlab-1.0.0.tar.gz" }
